feat(DRY) make codebase DRY  - create a reusable function to remove code that is being repeated.
