+++ graphql/resolvers/auth.js
@@ -1,0 +1,30 @@
+const bcrypt = require('bcryptjs');
+
+const User = require('../../models/user');
+
+
+module.exports = {
+  createUser: (args) => {
+    return User.findOne({ email: args.userInput.email })
+      .then((user) => {
+        if (user) {
+          throw new Error('User with that Email already exist');
+        }
+        return bcrypt.hash(args.userInput.password, 12);
+      })
+      .then((hashedPassword) => {
+        const user = new User({
+          username: args.userInput.username,
+          email: args.userInput.email,
+          password: hashedPassword,
+        });
+        return user.save();
+      })
+      .then((result) => {
+        return { ...result._doc, password: null, _id: result.id };
+      })
+      .catch((err) => {
+        throw err;
+      });
+  },
+};
