+++ helpers/index.js
@@ -1,0 +1,1 @@
+exports.convertRawDateToISO = rawDate => new Date(rawDate).toISOString();
